databases/work_with_database/requirements.txt
Django==2.1.1
python-slugify-3.0.3 
text-unidecode-1.2